Bugfix for empty POST with long body
Removed the fixed column and rows on the POST body
field as it was adding line breaks to the data going
through to sqlmap which was showing up as an empty
body.

The UI remains the same, this is due to `setLineWrap(True).

Bumped the version.

No more empty POSTs!